added registration and login functions
