--- frontend/src/App-temp.tsx
@@ -1,37 +1,0 @@
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Landing from "./Pages/Landing";
-import Pricing from "./Pages/Pricing/Pricing";
-import Registration from "./Pages/RegisterLogin/Registration";
-import Home from "./Pages/Learning/Home";
-import "./index.css";
-import { AuthProvider } from "./context/AuthContext";
-import PrivateRoute from "../components/routes/PrivateRoute";
-interface PrivateRouteProps {
-  component: React.ComponentType;
-}
-
-function App() {
-  return (
-    <Router>
-      <AuthProvider>
-        <div className="container px-6 mx-auto font-sans">
-          <Routes>
-            <Route path="/" element={<Landing />} />
-            <Route path="/pricing" element={<Pricing />} />
-            <Route path="/registration" element={<Registration />} />
-            <Route
-              path="/home"
-              element={
-                <PrivateRoute>
-                  <Home />
-                </PrivateRoute>
-              }
-            />
-          </Routes>
-        </div>
-      </AuthProvider>
-    </Router>
-  );
-}
-
-export default App;
